+++ spec/rails_helper.rb
@@ -2,12 +2,11 @@
 
 # Enabling code coverage reporting
 require 'simplecov'
-SimpleCov.start 'rails'
 
 # This file is copied to spec/ when you run 'rails generate rspec:install'
 require 'spec_helper'
 ENV['RAILS_ENV'] ||= 'test'
-require File.expand_path('dummy/config/environment.rb', __dir__)
+require File.expand_path(File.join(__dir__, 'dummy', 'config', 'environment.rb'))
 require 'rspec/rails'
 # Add additional requires below this line. Rails is not loaded until this point!
 
@@ -25,7 +24,7 @@ require 'rspec/rails'
 # require only the support files necessary.
 #
 
-Dir[File.join(__dir__, 'support', '**', '*.rb')].sort.each { |f| require f }
+Dir[File.join(__dir__, 'support', '**', '*.rb')].sort.each(&method(:require))
 
 # Checks for pending migrations and applies them before tests are run.
 # If you are not using ActiveRecord, you can remove these lines.